+++ src/components/API/__tests__/Integration/ExaremeLinearRegression.test.tsx
@@ -10,19 +10,25 @@ import {
   createModel,
   getDatasets,
   waitForResult
-} from '../../Utils';
+} from '../Utils';
 
 // config
 
 const modelSlug = `linear-${Math.round(Math.random() * 10000)}`;
-const experimentCode = 'LINEAR_REGRESSION';
+const experimentName = 'LINEAR_REGRESSION';
+const experimentLabel = 'Linear Regression'
 const parameters = [
   {
-    code: 'referencevalues',
+    name: 'referencevalues',
+    label: 'referencevalues',
     value: '[{"name":"alzheimerbroadcategory","val":"Other"}]'
   },
-  { code: 'encodingparameter', value: 'dummycoding' },
-  { code: 'pathology', value: 'dementia' }
+  {
+    name: 'encodingparameter',
+    label: 'encodingparameter',
+    value: 'dummycoding'
+  },
+  { name: 'pathology', label: 'pathology', value: 'dementia' }
 ];
 
 const model: any = (datasets: VariableEntity[]) => ({
@@ -70,16 +76,18 @@ describe('Integration Test for experiment API', () => {
     return datasets !== undefined && mstate.model !== undefined;
   });
 
-  it(`create ${experimentCode}`, async () => {
+  it(`create ${experimentName}`, async () => {
     if (!datasets) {
       throw new Error('datasets not defined');
     }
     const payload: ExperimentPayload = createExaremePayload(
       model,
       datasets,
-      experimentCode,
+      experimentName,
+      experimentLabel,
       parameters,
-      modelSlug
+      modelSlug,
+      'multiple_local_global'
     );
     const { error, experiment } = await createExperiment({
       experiment: payload
@@ -109,7 +117,7 @@ describe('Integration Test for experiment API', () => {
         .at(1)
         .first()
         .text()
-    ).toEqual('0.937');
+    ).toEqual('1.080');
     expect(
       wrapper
         .find('div.result')
@@ -118,6 +126,6 @@ describe('Integration Test for experiment API', () => {
         .at(1)
         .first()
         .text()
-    ).toEqual('-1.455');
+    ).toEqual('-0.942');
   });
 });