+++ src/components/API/__tests__/Integration/ExaremePearsonCorrelation.test.tsx
@@ -10,18 +10,19 @@ import {
   createModel,
   getDatasets,
   waitForResult
-} from '../../Utils';
+} from '../Utils';
 
 // config
 
 const modelSlug = `pearson-${Math.round(Math.random() * 10000)}`;
-const experimentCode = 'PEARSON_CORRELATION';
+const experimentName = 'PEARSON_CORRELATION';
+const experimentLabel = 'Pearson Correlation';
 const parameters = [
-  { code: 'bins', value: '40' },
-  { code: 'iterations_max_number', value: 20 },
-  { code: 'sstype', value: 2 },
-  { code: 'outputformat', value: 'pfa' },
-  { code: 'pathology', value: 'dementia' }
+  { name: 'bins', value: '40', label: 'bins' },
+  { name: 'iterations_max_number', value: 20, label: 'iterations_max_number' },
+  { name: 'sstype', value: 2, label: 'sstype' },
+  { name: 'outputformat', value: 'pfa', label: 'outputformat' },
+  { name: 'pathology', value: 'dementia', label: 'pathology' }
 ];
 
 const model: any = (datasets: VariableEntity[]) => ({
@@ -68,16 +69,18 @@ describe('Integration Test for experiment API', () => {
     return datasets !== undefined && mstate.model !== undefined;
   });
 
-  it(`create ${experimentCode}`, async () => {
+  it(`create ${experimentName}`, async () => {
     if (!datasets) {
       throw new Error('datasets not defined');
     }
     const payload: ExperimentPayload = createExaremePayload(
       model,
       datasets,
-      experimentCode,
+      experimentName,
+      experimentLabel,
       parameters,
-      modelSlug
+      modelSlug,
+      'python_local_global'
     );
     const { error, experiment } = await createExperiment({
       experiment: payload