+++ src/components/App/App.tsx
@@ -1,16 +1,15 @@
 import 'bootstrap/dist/css/bootstrap.css';
-
 import * as React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import styled, { createGlobalStyle } from 'styled-components';
-import { ExperimentResponse } from '../API/Experiment';
 import backgroundImage from '../../images/body-bg.jpg';
 import ExperimentReview from '../Analysis/Container';
 import { APICore, APIExperiment, APIMining, APIModel, APIUser } from '../API';
+import { ExperimentResponse } from '../API/Experiment';
 import Article from '../Article/Container';
 import ExperimentCreate from '../Create/Container';
-import Home from '../Dashboard/Container';
 import Explore from '../Explore/Container';
+import Help from '../Help/Help';
 import ExperimentResult from '../Result/Container';
 import Footer from '../UI/Footer';
 import Galaxy from '../UI/Galaxy';
@@ -19,14 +18,10 @@ import NotFound from '../UI/NotFound';
 import TOS from '../UI/TOS';
 import User from '../User/Container';
 import { history } from '../utils';
-export enum InstanceMode {
-  Local,
-  Federation
-}
+
 export interface AppConfig {
   version?: string;
   instanceName?: string;
-  mode?: InstanceMode;
   ga?: string;
 }
 interface Props {
@@ -40,10 +35,9 @@ interface Props {
 
 const GlobalStyles = createGlobalStyle`
   body {
-    font-family: 'Open Sans', sans-serif;
+    font-family: 'Open Sans', sans-serif !important;
     background: url(${backgroundImage}) top center no-repeat fixed #f5f5f5;
     background-size: 100% auto;
-    overflow-y: scroll
   }
 
   .panel {
@@ -53,12 +47,12 @@ const GlobalStyles = createGlobalStyle`
     border-radius: 4px;
     box-shadow: 0 1px 1px rgba(0, 0, 0, 0.05);
   }
-}
 `;
 
 const Main = styled.main`
-  margin: 52px 0 32px 0;
-  padding: 0 48px 0px 48px;
+  margin: 0;
+  padding: 52px 8px;
+  min-height: 100vh;
 `;
 
 const App = ({
@@ -78,37 +72,16 @@ const App = ({
           experiments={apiExperiment.state.experiments}
           handleSelect={(experiment: ExperimentResponse): void => {
             history.push(
-              `/experiment/${experiment.modelDefinitionId}/${experiment.uuid}`
+              `/experiment/${experiment.modelSlug}/${experiment.uuid}`
             );
           }}
-          isFederated={appConfig.mode === InstanceMode.Federation}
         />
       </header>
       <Main>
         <Switch>
           <Route
-            path="/"
+            path={['/', '/explore']}
             exact={true}
-            // tslint:disable-next-line jsx-no-lambda
-            render={(props): JSX.Element => (
-              <Home
-                apiCore={apiCore}
-                apiModel={apiModel}
-                apiExperiment={apiExperiment}
-                apiUser={apiUser}
-                {...props}
-              />
-            )}
-          />
-          <Route
-            path="/tos"
-            // tslint:disable-next-line jsx-no-lambda
-            render={(props): JSX.Element => (
-              <TOS apiUser={apiUser} {...props} />
-            )}
-          />
-          <Route
-            path="/explore"
             // tslint:disable-next-line jsx-no-lambda
             render={(props): JSX.Element => (
               <Explore
@@ -120,6 +93,14 @@ const App = ({
               />
             )}
           />
+          <Route
+            path="/tos"
+            // tslint:disable-next-line jsx-no-lambda
+            render={(props): JSX.Element => (
+              <TOS apiUser={apiUser} {...props} />
+            )}
+          />
+
           <Route
             path="/review"
             // tslint:disable-next-line jsx-no-lambda
@@ -174,6 +155,12 @@ const App = ({
             render={(props): JSX.Element => (
               <User apiUser={apiUser} {...props} />
             )}
+          />
+
+          <Route
+            path="/training"
+            // tslint:disable-next-line jsx-no-lambda
+            render={(props): JSX.Element => <Help />}
           />
 
           <Route component={NotFound} />