+++ src/components/Result/RenderResult.tsx
@@ -2,10 +2,11 @@ import * as React from 'react';
 
 import { Result } from '../API/Experiment';
 import { MIME_TYPES } from '../constants';
-import { Highchart, JSONData } from '../UI/Visualization';
 import Error from '../UI/Error';
-import Warning from '../UI/Visualization/Warning';
+import { Highchart, JSONData } from '../UI/Visualization';
 import Dendogram from '../UI/Visualization/Dendogram';
+import BinaryTree from '../UI/Visualization/BinaryTree';
+import Warning from '../UI/Visualization/Warning';
 
 export default ({
   results
@@ -37,6 +38,9 @@ export default ({
                 )}
                 {result.type === MIME_TYPES.HIGHCHARTS && (
                   <Highchart options={result.data} />
+                )}
+                {result.type === MIME_TYPES.JSONBTREE && (
+                  <BinaryTree data={result.data} />
                 )}
                 {result.type === MIME_TYPES.JSON && (
                   <Dendogram data={result.data} />