+++ src/components/constants.tsx
@@ -5,91 +5,107 @@ export enum MIME_TYPES {
   USER_WARNING = 'text/plain+user_error',
   HIGHCHARTS = 'application/vnd.highcharts+json',
   JSON = 'application/json',
-  MIP_PFA = 'application/vnd.hbp.mip.experiment.pfa+json',
-  MIP_COMPOUND = 'application/vnd.hbp.mip.compound+json',
+  JSONBTREE = 'application/binary-tree+json',
   PFA = 'application/pfa+json',
-  PLOTLY = 'application/vnd.plotly.v1+json',
   JSONDATA = 'application/vnd.dataresource+json',
   HTML = 'text/html',
-  TEXT = 'text/plain',
-  JSONRAW = 'application/raw+json'
+  TEXT = 'text/plain'
 }
 
-const DEFAULT_OUTPUT = [
+export const HISTOGRAMS_STORAGE_KEY = 'mipChoosenHistogramsVars';
+
+export const ERRORS_OUTPUT = [
   MIME_TYPES.ERROR,
   MIME_TYPES.WARNING,
   MIME_TYPES.USER_WARNING
 ];
 
-// TODO: merge OUTPUT and ENABLED
-export const ALGORITHM_DEFAULT_OUTPUT = [
-  {
-    name: 'ANOVA',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA]
-  },
-  {
-    name: 'LINEAR_REGRESSION',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA]
-  },
-  {
-    name: 'LOGISTIC_REGRESSION',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'HISTOGRAMS',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'TTEST_INDEPENDENT',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA]
-  },
-  {
-    name: 'TTEST_PAIRED',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA]
-  },
-  {
-    name: 'PEARSON_CORRELATION',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'ID3',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.JSON]
-  },
-  {
-    name: 'KMEANS',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'NAIVE_BAYES_TRAINING_STANDALONE',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'TTEST_ONESAMPLE',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA]
-  },
-  {
-    name: 'MULTIPLE_HISTOGRAMS',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.HIGHCHARTS]
-  },
-  {
-    name: 'PCA',
-    types: [...DEFAULT_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
-  }
-];
-
 export const ENABLED_ALGORITHMS = [
-  'LOGISTIC_REGRESSION',
-  'ANOVA',
-  'NAIVE_BAYES_TRAINING_STANDALONE',
-  'KMEANS',
-  'PEARSON_CORRELATION',
-  'ID3',
-  // 'HISTOGRAMS',
-  'LINEAR_REGRESSION',
-  'TTEST_INDEPENDENT',
-  'TTEST_PAIRED',
-  'TTEST_ONESAMPLE',
-  'PCA'
+  {
+    enabled: true,
+    label: 'ANOVA',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA]
+  },
+  {
+    enabled: true,
+    label: 'Linear Regression',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA]
+  },
+  {
+    enabled: true,
+    label: 'Logistic Regression',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: false,
+    label: 'Histograms',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'T-Test Independent',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA]
+  },
+  {
+    enabled: true,
+    label: 'T-Test Paired',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA]
+  },
+  {
+    enabled: true,
+    label: 'Pearson Correlation',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'ID3',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.JSON]
+  },
+  {
+    enabled: true,
+    label: 'k-Means Clustering',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'Naive Bayes Training',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'T-Test One-Sample ',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA]
+  },
+  {
+    enabled: false,
+    label: 'Multiple Histograms',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'Principal Components analysis',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSONDATA, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'Workflow Naive Bayes with Hold Out Validation',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'Naive Bayes with Cross Validation',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'Calibration Belt',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.HIGHCHARTS]
+  },
+  {
+    enabled: true,
+    label: 'CART',
+    types: [...ERRORS_OUTPUT, MIME_TYPES.JSON]
+  }
 ];
 
 const independents = ['X', 'column1', 'x', 'descriptive_attributes'];