+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+### dev_5.2.4 - 10.03.2020
+
+- Cart algorithm integrated
+- Updated Typescript & d3.js librairies
+- Histogram grouping variables fixes
+- Annotation module for Highchart
+- Changed TBI CDEs Mortality to Categorical type
+- hbpmip/exareme:dev_v14.2
+- Calibration Belt dynamic variables integration
+- Added Annotation module for Highcharts
+- Reverted x<->y for multiple histograms
+
+### 5.2.0 - 27.02.2020
+
+- New API for Galaxy integration
+- Dockerized tests - (run-test.sh)
+- Calibration Belt integration
+- Removed hint on algorithm hover
+- Histogram fixes
+- Removed local/federated mode
+- Added training videos in help
+- Bug fixes
+- UI improvements
+
 ### 5.1.18 - 30.01.2020
 
 - Added more static filters for dementia